+++ modal-go/cls.go
@@ -27,7 +27,12 @@ func ClsLookup(ctx context.Context, appName string, name string, options *Lookup
 	if options == nil {
 		options = &LookupOptions{}
 	}
-	ctx = clientContext(ctx)
+	var err error
+	ctx, err = clientContext(ctx)
+	if err != nil {
+		return nil, err
+	}
+
 	cls := Cls{
 		methodNames: []string{},
 		ctx:         ctx,