+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Both client libraries are pre-1.0, and they have separate versioning.
 
 ## Unreleased
 
-No unreleased changes.
+- Add `InitializeClient()` (Go) / `initializeClient()` (TS) to initialize the client at runtime with credentials.
+- Client libraries no longer panic at startup if no token ID / secret is provided. Instead, they will throw an error when trying to use the client.
 
 ## modal-js/v0.3.10, modal-go/v0.0.10